create calendar element
