+++ front/src/components/CalendarBoard/style.css
@@ -1,11 +1,7 @@
 .container {
   height: 90vh;
 }
-.element {
-  border-right: 1px solid #ccc;
-  border-bottom: 1px solid #ccc;
-  height: 18vh;
-}
+
 .grid {
   border-left: 1px solid #ccc;
   border-top: 1px solid #ccc;
